Move Multi Tab Tests to own file and re-enable one multi tab test The multi tab test was marked as fix me, re-enabling now as we need a fix for this
