+++ Test/End-2-End/scenario_testing.spec.js
@@ -38,111 +38,6 @@ test("Can place an order using different billing and shipping address", async ({
   await new OrderConfirmation(page).expectOnOrderConfirmation();
 });
 
-test.describe("multiple tabs", () => {
-  test("Changing quote on a different tab for an in progress order makes the payment invalid", async ({
-    browser,
-  }) => {
-    const context = await browser.newContext();
-    const mainPage = await context.newPage();
-    const visitCheckoutPayment = new VisitCheckoutPayment(mainPage);
-    await visitCheckoutPayment.visit();
-    await mainPage.getByLabel("Rvvup Payment Method").click();
-    await mainPage.getByRole("button", { name: "Place order" }).click();
-    const frame = mainPage.frameLocator(
-      "#rvvup_iframe-rvvup_FAKE_PAYMENT_METHOD",
-    );
-    await frame.getByRole("button", { name: "Pay now" }).isVisible();
-
-    const duplicatePage = await context.newPage();
-    const cart = new Cart(duplicatePage);
-    await cart.addStandardItemToCart("Rvvup Crypto Future");
-
-    await frame.getByRole("button", { name: "Pay now" }).click();
-
-    await mainPage.waitForURL("**/checkout/cart/");
-    await expect(
-      mainPage.getByText(
-        "Your cart was modified after making payment request, please place order again.",
-      ),
-    ).toBeVisible();
-  });
-
-  test.fixme(
-    "Cannot complete the same order in multiple tabs simultaneously",
-    async ({ browser }) => {
-      const context = await browser.newContext();
-
-      // Start the checkout on the first tab
-      const mainPage = await context.newPage();
-      await new VisitCheckoutPayment(mainPage).visit();
-
-      await mainPage.getByLabel("Rvvup Payment Method").click();
-      await mainPage.getByRole("button", { name: "Place order" }).click();
-      const mainFrame = mainPage.frameLocator(
-        "#rvvup_iframe-rvvup_FAKE_PAYMENT_METHOD",
-      );
-
-      // Start the checkout on the second tab
-      const duplicatePage = await context.newPage();
-      await new GoTo(duplicatePage).checkout();
-
-      await duplicatePage.getByRole("button", { name: "Next" }).click();
-
-      await duplicatePage.getByLabel("Rvvup Payment Method").click();
-      await duplicatePage.getByRole("button", { name: "Place order" }).click();
-      const duplicateFrame = duplicatePage.frameLocator(
-        "#rvvup_iframe-rvvup_FAKE_PAYMENT_METHOD",
-      );
-
-      // Complete order in the first tab, and then in the second tab shortly after
-      await duplicateFrame.getByRole("button", { name: "Pay now" }).click();
-
-      await new OrderConfirmation(duplicatePage).expectOnOrderConfirmation();
-
-      await mainFrame.getByRole("button", { name: "Pay now" }).click();
-      await expect(
-        mainPage.getByText(/An error has happened during application run.*/),
-      ).toBeVisible();
-    },
-  );
-
-  test("Cannot place order in one tab and then place the same order again in another tab", async ({
-    browser,
-  }) => {
-    const context = await browser.newContext();
-
-    // Start the checkout on the first tab
-    const mainPage = await context.newPage();
-    await new VisitCheckoutPayment(mainPage).visit();
-    await mainPage.getByLabel("Rvvup Payment Method").click();
-
-    // Open the checkout page on the second tab
-    const duplicatePage = await context.newPage();
-    await new GoTo(duplicatePage).checkout();
-
-    await duplicatePage.getByRole("button", { name: "Next" }).click();
-    await expect(
-      duplicatePage.getByText("Payment Method", { exact: true }),
-    ).toBeVisible();
-    await duplicatePage.getByLabel("Rvvup Payment Method").click();
-
-    // Complete order in the first tab
-    await mainPage.getByRole("button", { name: "Place order" }).click();
-    const mainFrame = mainPage.frameLocator(
-      "#rvvup_iframe-rvvup_FAKE_PAYMENT_METHOD",
-    );
-    await mainFrame.getByRole("button", { name: "Pay now" }).click();
-
-    await new OrderConfirmation(mainPage).expectOnOrderConfirmation();
-
-    // Complete order in the second tab
-    await duplicatePage.getByRole("button", { name: "Place order" }).click();
-    await expect(
-      duplicatePage.getByText("No such entity with cartId ="),
-    ).toBeVisible();
-  });
-});
-
 test.describe("discounts", () => {
   test("Can place an order using discount codes", async ({ page }) => {
     await new VisitCheckoutPayment(page).visit();
